pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rfq-e2e-tests"
version = "0.1.0"
description = "E2E and Load Testing Framework for RFQ System"
requires-python = ">=3.11"

dependencies = [
    "injective-py>=1.0.0",
    "websockets>=12.0",
    "eth-account>=0.11.0",
    "eth-hash[pycryptodome]>=0.5.0",
    "bech32>=1.2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-xdist>=3.5",
    "pyyaml>=6.0",
    "rich>=13.0",
    "allure-pytest>=2.13",
    "tenacity>=8.2",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0",
    "types-PyYAML",
]

[tool.hatch.build.targets.wheel]
packages = ["src/rfq_test"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
norecursedirs = ["archive"]
markers = [
    "smoke: Quick health check tests (<30s total)",
    "slow: Long-running tests",
    "load: Load/stress tests",
    "validation: Input validation tests",
    "functional: E2E flow tests",
    "contract: Smart contract validation tests",
    "local_only: Tests that only work on local chain",
    "setup: Setup tests (whitelist makers, etc.) - run before other tests",
]
addopts = "-v --tb=short"
filterwarnings = [
    # pyinjective AsyncClient teardown uses deprecated asyncio.get_event_loop() when pytest loop is already closed
    "ignore::DeprecationWarning:pyinjective.async_client_v2",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
